Spotbugs fix.
List of removed metrics:

- kafka_poll_request
- kafka_poll_request_latency (was not being populated)
- hybrid_quota_enforcement_latency
- consumer_to_queue_latency
- produce_to_drainer_queue_call_count
- produce_to_drainer_queue_record_num
- produce_to_kafka_record_num (at least 1 other metric also reports this)
- produce_to_kafka_latency
- data_validation (DIV latency)

+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/stats/LongAdderRateGauge.java
@@ -20,8 +20,8 @@ public class LongAdderRateGauge extends Gauge {
 
   @Override
   public double measure(MetricConfig config, long now) {
-    long elapsedTimeInSeconds = (now - this.lastMeasurementTime) / Time.MS_PER_SECOND;
+    double elapsedTimeInSeconds = (double) (now - this.lastMeasurementTime) / Time.MS_PER_SECOND;
     this.lastMeasurementTime = now;
-    return this.adder.sumThenReset() / elapsedTimeInSeconds;
+    return (double) this.adder.sumThenReset() / elapsedTimeInSeconds;
   }
 }
